+++ apps/fluent-tester/metro.config.js
@@ -1,12 +1,5 @@
-/**
- * Metro configuration for React Native
- * https://github.com/facebook/react-native
- *
- * @format
- */
-
 const path = require('node:path');
-const { exclusionList, makeMetroConfig, resolveUniqueModule } = require('@rnx-kit/metro-config');
+const { makeMetroConfig, exclusionList } = require('@rnx-kit/metro-config');
 const MetroSymlinksResolver = require('@rnx-kit/metro-resolver-symlinks');
 
 // ensure regex paths are merged, normalized, use forward slashes and end with a /
@@ -18,38 +11,27 @@ function pathForRegex(...parts) {
   return result.replace(/[/\\]+/g, '/');
 }
 
-const excludeMixins = [];
-const extraNodeModules = {};
-
-function ensureUniqueModule(moduleName) {
-  const [nmEntry, excludePattern] = resolveUniqueModule(moduleName);
-  excludeMixins.push(excludePattern);
-  extraNodeModules[moduleName] = nmEntry;
-}
-
-// build up the added excludes and extraNodeModules
-['react-native-svg'].forEach((moduleName) => ensureUniqueModule(moduleName));
-
 const blockList = exclusionList([
   // Exclude other test apps
   new RegExp(pathForRegex(__dirname, '../win32')),
-
+  new RegExp(pathForRegex(__dirname, '../win32-81')),
   // Exclude build output directory
   new RegExp(pathForRegex(__dirname, 'dist')),
-
-  ...excludeMixins,
 ]);
 
-let config = makeMetroConfig({
+const config = makeMetroConfig({
   resolver: {
     blockList,
-    extraNodeModules,
     resolveRequest: MetroSymlinksResolver({
       resolver: 'oxc-resolver',
     }),
+    unstable_enablePackageExports: true,
+    unstable_conditionNames: ['react-native', 'import', 'require'],
+
+    disableHierarchicalLookup: true,
+    enableSymlinks: true,
   },
   transformer: {
-    // This transformer selects between the regular transformer and svg transformer depending on the file type
     babelTransformerPath: require.resolve('react-native-svg-transformer'),
   },
 });