+++ packages/experimental/Dropdown/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log - @fluentui-react-native/dropdown
 
+## 0.11.4
+
+### Patch Changes
+
+- ce37a6d: Remove the final configuration package (moving it to scripts) and switch to running various package validation tasks from the root
+- ad85949: Configuration and dependency changes
+- 51ee2fe: Bump win32 to 0.81
+- f10715d: remove oxlint config package and move it into scripts
+- 1598157: Switch tsconfig settings to more modern settings
+- Updated dependencies [ce37a6d]
+- Updated dependencies [b5a686c]
+- Updated dependencies [ad85949]
+- Updated dependencies [23aa900]
+- Updated dependencies [51ee2fe]
+- Updated dependencies [f10715d]
+- Updated dependencies [1598157]
+  - @fluentui-react-native/interactive-hooks@0.28.4
+  - @fluentui-react-native/theme-tokens@0.28.4
+  - @fluentui-react-native/framework@0.16.0
+  - @fluentui-react-native/callout@0.28.4
+  - @fluentui-react-native/button@0.40.4
+  - @fluentui-react-native/text@0.25.4
+  - @fluentui-react-native/adapters@0.14.4
+
 ## 0.11.3
 
 ### Patch Changes