+++ packages/utils/styling/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log - @fluentui-react-native/styling-utils
 
+## 0.8.4
+
+### Patch Changes
+
+- ce37a6d: Remove the final configuration package (moving it to scripts) and switch to running various package validation tasks from the root
+- ad85949: Configuration and dependency changes
+- f10715d: remove oxlint config package and move it into scripts
+- 1598157: Switch tsconfig settings to more modern settings
+
 ## 0.8.3
 
 ### Patch Changes